Extract out `validateBlueprintData` logic
